add base brownie console script
