+++ brownie/addresses.py
@@ -59,6 +59,16 @@ OETH_LIDO_WITHDRAWAL_STRAT = '0xD9B488280d723338Dd32d56b3900f379Eb7A7aF1'
 VAULT_OETHB_PROXY_ADDRESS = '0x98a0CbeF61bD2D21435f433bE4CD42B56B38CC93'
 OETHB = '0xDBFeFD2e8460a6Ee4955A68582F85708BAEA60A3'
 
+# Base contracts
+WETH_BASE = '0x4200000000000000000000000000000000000006'
+AERO_BASE = '0x940181a94A35A4569E4529A3CDfB74e38FD98631'
+AERODROME_SWAP_ROUTER_BASE = '0xBE6D8f0d05cC4be24d5167a3eF062215bE6D18a5'
+AERODROME_POSITION_MANAGER_BASE = '0x827922686190790b37229fd06084350E74485b72'
+AERODROME_WETH_OETHB_POOL_BASE = '0x6446021F4E396dA3df4235C62537431372195D38'
+AERODROME_WETH_OETHB_GAUGE_BASE = '0x6446021F4E396dA3df4235C62537431372195D38'
+AERODROME_SUGAR_HELPER_BASE = '0x0AD09A66af0154a84e86F761313d02d0abB6edd5'
+
+
 # OGV Contracts
 OGV = "0x9c354503C38481a7A7a51629142963F98eCC12D0"
 REWARDS = "0x7d82E86CF1496f9485a8ea04012afeb3C7489397"
@@ -102,7 +112,6 @@ UNISWAP_V3_QUOTER = '0x61fFE014bA17989E743c5F6cB21bF9697530B21e'
 OETH = "0x856c4efb76c1d1ae02e20ceb03a2a6a08b0b8dc3"
 OETH_ZAPPER = "0x9858e47BCbBe6fBAC040519B02d7cd4B2C470C66"
 
-WETH = "0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2"
 FRXETH = "0x5e8422345238f34275888049021821e8e08caa1f"
 SFRXETH = "0xac3e018457b222d93114458476f3e3416abbe38f"
 RETH = "0xae78736cd615f374d3085123a210448e74fc6393"
